feat: keep chunk size in header
