+++ src/core/encryption/ciphers/Twofish/index.ts
@@ -1,55 +1,19 @@
 import { TwofishModule } from 'twofish';
 import twofishModule from 'twofish/twofish.wasm';
-import { bytes, struct, u8 } from '@core/encryption/utils/bytes/binstruct';
+import { bytes, struct, u8, u32 } from '@core/encryption/utils/bytes/binstruct';
 import { xor16 } from '@core/encryption/utils/xor';
 
 import { CTRCipherMode } from '../../cipherModes/CTRCipherMode';
 import { fillBuffer } from '../../utils/buffers';
+import { HKDFDerivedKeys } from '../../utils/HKDFDerivedKeys';
 
 import { IEncryptionProcessor } from '../..';
-
-export class HKDFDerivedKeys {
-	private readonly key;
-	constructor(
-		key: Uint8Array<ArrayBuffer>,
-		private readonly iv: Uint8Array<ArrayBuffer>,
-	) {
-		this.key = crypto.subtle.importKey('raw', key, 'HKDF', false, [
-			'deriveKey',
-			'deriveBits',
-		]);
-	}
-
-	public stringToBuffer(text: string) {
-		return new TextEncoder().encode(text);
-	}
-
-	public async deriveBits(length: number, context: string | Uint8Array<ArrayBuffer>) {
-		const key = await this.key;
-
-		return await crypto.subtle.deriveBits(
-			{
-				name: 'HKDF',
-				hash: 'SHA-256',
-				salt: this.iv,
-				info:
-					typeof context === 'string' ? this.stringToBuffer(context) : context,
-			},
-			key,
-			length,
-		);
-	}
-
-	public async deriveBytes(length: number, context: string | Uint8Array<ArrayBuffer>) {
-		const buffer = await this.deriveBits(length * 8, context);
-		return new Uint8Array(buffer);
-	}
-}
 
 export const TWOFISH_IV_SIZE = 12;
 export const MASTER_IV_SIZE = 24;
 export const TWOFISH_HEADER = struct({
 	padding: u8(),
+	chunkSize: u32(),
 	iv: bytes(MASTER_IV_SIZE),
 });
 
@@ -95,7 +59,7 @@ export class WasmTwofishCTRCipher implements IEncryptionProcessor {
 		const output = new Uint8Array(
 			new ArrayBuffer(TWOFISH_HEADER.size + input.length),
 		);
-		output.set(TWOFISH_HEADER.encode({ padding, iv }), 0);
+		output.set(TWOFISH_HEADER.encode({ padding, iv, chunkSize: this.chunkSize }), 0);
 		const encryptedData = output.subarray(TWOFISH_HEADER.size);
 
 		const derivedKey = await hkdf.deriveBytes(16, `key`);
@@ -125,7 +89,7 @@ export class WasmTwofishCTRCipher implements IEncryptionProcessor {
 	}
 
 	public async decrypt(buffer: ArrayBuffer) {
-		const { padding, iv } = TWOFISH_HEADER.decode(
+		const { padding, iv, chunkSize } = TWOFISH_HEADER.decode(
 			new Uint8Array(buffer, 0, TWOFISH_HEADER.size),
 		);
 
@@ -141,7 +105,7 @@ export class WasmTwofishCTRCipher implements IEncryptionProcessor {
 		const input = new Uint8Array(buffer, TWOFISH_HEADER.size);
 		const output = new Uint8Array(input.length - padding);
 		const derivedKey = await hkdf.deriveBytes(16, `key`);
-		const chunksCount = Math.ceil(input.byteLength / this.chunkSize);
+		const chunksCount = Math.ceil(input.byteLength / chunkSize);
 		for (let i = 0; i < chunksCount; i++) {
 			const derivedNonce = await hkdf.deriveBytes(TWOFISH_IV_SIZE, `nonce${i}`);
 
@@ -152,9 +116,9 @@ export class WasmTwofishCTRCipher implements IEncryptionProcessor {
 					(a, b) => xor16(xorBuffer, a, b),
 				);
 
-				const offset = this.chunkSize * i;
+				const offset = chunkSize * i;
 				const decryptedChunk = await cipher.decrypt(
-					input.subarray(offset, offset + this.chunkSize),
+					input.subarray(offset, offset + chunkSize),
 					derivedNonce,
 				);
